requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
Django==3.0.7
djangorestframework==3.11.0
fake-useragent==0.1.11
idna==2.9
lxml==4.5.1
parsel==1.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
selectorlib==0.16.0
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
w3lib==1.22.0
